MT#60205 Port lua-ngcp-kamailio and templates logic and create new API routes
- Bundles existing logic from the lua-ngcp-kamailio
and template modules for dialog counter manipulation
into the RedisManager and DialogManager classes.

- Exposes several API endpoints to interact with the
public methods of the DialogManager class for writing
and editing/deleting counters on the redit databases

- Adds new configuration for the redis database
cluster with a central database housing dialog
counters and a local database where the dialog ids are stored.

- Fixes routes not included in API and redis servers
not being correctly exposed.

- Adds api documentation with examples and
 descriptors for every endpoint. Additionally, some minor
 bug fixes were implemented

Change-Id: Ida9fd1777454cf4836fae6140b1aa6db1d98fa61